Updated Contact form
Now depends on MailJS
Requires a valid email
Requires a non-null name, subject and message.
The error message will be shown if something is wrong.
If an mail is sent then it will be displayed.
It is also displayed to the user if something is loading.
Made changes to experiences.